Fix retrieving source files logic
