+++ components/context/src/main/resources/META-INF/services/org.junit.jupiter.api.extension.Extension
@@ -1,1 +1,1 @@
-datadog.context.TestSourceFileExtension
+datadog.context.GatherSourceFileInfoExtension
